backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
mysql-connector-python==8.3.0
celery==5.3.6
redis==5.0.1
minio==7.2.3
qdrant-client==1.7.3
openai==1.10.0
python-dotenv==1.0.0
PyPDF2==3.0.1
python-docx==1.1.0
sentence-transformers==2.3.1
